Refactor telemetry pipeline, upgrade database precision, and switch to DotNetEnv for configuration

+++ server/Api/Entities/FermentationTelemetryEntity.cs
@@ -5,8 +5,8 @@ namespace Api.Entities
         public Guid Id { get; set; }
         public string DeviceId { get; set; } = string.Empty;
         public bool FermentationActive { get; set; }
-        public float Temperature { get; set; }
-        public float EthanolSignal { get; set; }
+        public double Temperature { get; set; }
+        public double EthanolSignal { get; set; }
         public DateTime Timestamp { get; set; } 
     }
 }
